requirements.txt
llama-cpp-python
fastapi
uvicorn
streamlit
st-annotated-text
pytest
